Gross blinky
A basic blinky program where an LED is turned on and off repeatably. The
project runs on raytac's development [board](
https://www.raytac.com/product/ins.php?index_id=114) for their nr52805
[module](https://www.raytac.com/product/ins.php?index_id=108).

This implementation is gross because it uses an empty for loop to delay
between turning the LED on and off. This implementation is dependent on
the processor speed and the compiler's optimization level.